Move commands into own subpackage
