+++ lammps_testing/lammps_test.py
@@ -1,103 +1,20 @@
 #!/usr/bin/python
 import argparse
 import glob
-import logging
-import math
 import os
-import platform
 import re
-import shutil
 import subprocess
 import sys
-import threading
-import time
-import json
-from datetime import datetime
-from pathlib import Path
-from collections import namedtuple
-
-import yaml
 
 from termcolor import colored
-from .common import logger, Settings, Container, discover_tests
+from .common import logger, Settings, Container, discover_tests, get_containers, get_configurations, get_container, get_configuration, get_names
+from .tests import CompilationTest, RunTest, RegressionTest, UnitTest
 
+from .commands.env import init_command as init_env_command
+from .commands.config import init_command as init_config_command
+from .commands.build import init_command as init_build_command
+from .commands.reg import init_command as init_reg_command
 
-class LAMMPSExampleRun(object):
-    def __init__(self, name, input_script, configuration):
-        self.name = name
-        self.input_script = input_script
-        self.configuration = configuration
-
-    def __str__(self):
-        if "mpi" in self.configuration:
-            return f"{self.name} (MPI: {self.configuration['mpi']['nprocs']} procs)"
-        return f"{self.name} (Serial)"
-
-
-def config_from_logfile(logfile):
-    """This converts the logfile filename into a configuration"""
-    parts = os.path.basename(logfile).split('.')
-    ref_date =  parts[1]
-    compiler = parts[-2]
-    nprocs = int(parts[-1])
-
-    config = {
-        "reference" : {
-            "logfile": logfile,
-            "date": ref_date
-        }
-    }
-
-    config['compiler'] = compiler
-
-    if nprocs > 1:
-        config['mpi'] = {'nprocs': nprocs}
-
-    return config
-
-class LAMMPSExample(object):
-    def __init__(self, name, scripts, logfiles):
-        self.name = name
-        self.testcases = {}
-        self.folder = os.path.dirname(scripts[0])
-
-        for script in scripts:
-            script_name = os.path.basename(script)
-            testcase_name = script_name[3:]
-            logfile_pattern = re.compile(r"log.(?P<date>[0-9]+[a-zA-Z]{3}[0-9]{2})\." + testcase_name + r"\.(?P<config>(g\+\+|clang)\.[1-9]*)")
-            run = LAMMPSExampleRun(testcase_name, script, {'mpi': {'nprocs': 8}})
-            self.testcases[testcase_name] = [run]
-
-            for logfile in logfiles:
-                m = logfile_pattern.match(os.path.basename(logfile))
-                if m is not None:
-                    try:
-                        configuration = config_from_logfile(logfile)
-                        run = LAMMPSExampleRun(testcase_name, script, configuration)
-                        self.testcases[testcase_name].append(run)
-                    except ValueError:
-                        pass
-
-    def save_config(self):
-        config_file = os.path.join(self.folder, '.testing', 'config.yaml')
-        config = []
-
-        for testcase_name, runs in self.testcases.items():
-            run_list = []
-
-            for run in runs:
-                logfile = os.path.basename(run.configuration['logfile'])
-                if 'mpi' in run.configuration:
-                    run_list.append({'logfile': logfile, 'mpi': {'nprocs': run.configuration['mpi']['nprocs']}})
-                else:
-                    run_list.append({'logfile': logfile})
-
-            testcase = {"input_script": f"in.{testcase_name}", "runs": run_list}
-            config.append(testcase)
-
-        os.makedirs(os.path.dirname(config_file),exist_ok=True)
-        with open(config_file, "w") as f:
-            yaml.dump(config, f)
 
 
 class TestRunner():
@@ -186,54 +103,7 @@ class MPIRunner(LocalRunner):
         env["LAMMPS_LAUNCHER"] = f'mpirun -np {self.nprocs} '
         return env
 
-def get_container(name, settings):
-    if name == 'local':
-        return None
-    container = os.path.join(settings.container_dir, name + ".sif")
-    container_definition = os.path.join(settings.container_definition_dir, name + ".def")
-    return Container(name, container, container_definition)
 
-def get_lammps_commit(commit, settings):
-    return subprocess.check_output(['git', 'rev-parse', commit], cwd=settings.lammps_dir).decode().strip()
-
-def get_names(search_pattern):
-    names = []
-    for path in glob.glob(search_pattern):
-        base = os.path.basename(path)
-        name = os.path.splitext(base)[0]
-        names.append(name)
-    return names
-
-def get_containers(settings):
-    containers  = get_names(os.path.join(settings.container_definition_dir, '*.def'))
-    containers += get_names(os.path.join(settings.container_definition_dir, '**/*.def'))
-    return [get_container(c, settings) for c in sorted(containers)]
-
-def get_containers_by_selector(selector, settings):
-    if 'all' in selector or 'ALL' in selector:
-        return get_containers(settings)
-    return [get_container(c, settings) for c in sorted(selector)]
-
-def get_configurations(settings):
-    configurations = get_names(os.path.join(settings.configuration_dir, '*.yml'))
-    return [get_configuration(c, settings) for c in sorted(configurations)]
-
-def get_commits(settings):
-    commits = get_names(os.path.join(settings.cache_dir, 'builds_*'))
-    return [c[7:] for c in sorted(commits)]
-
-def get_configuration(name, settings):
-    configfile = os.path.join(settings.configuration_dir, name + ".yml")
-    if os.path.exists(configfile):
-        with open(configfile) as f:
-            config = yaml.full_load(f)
-            return namedtuple("Configuration", ['name'] + list(config.keys()))(name, *config.values())
-    raise FileNotFoundError(configfile)
-
-def get_configurations_by_selector(selector, settings):
-    if 'all' in selector or 'ALL' in selector:
-        return get_configurations(settings)
-    return [get_configuration(c, settings) for c in sorted(selector)]
 
 def get_lammps_runner(runner, builder, settings):
     if runner == 'testing':
@@ -292,28 +162,6 @@ def status(args, settings):
             for regtest in config.regression_tests:
                 print(" ", f"{regtest:<40}")
 
-def env_clean(args, settings):
-    containers = get_containers_by_selector(args.images, settings)
-
-    print("Selected Containers:")
-
-    for container in containers:
-        print("-", container.name)
-
-    if not args.force:
-        answer = input('Delete all these containers? (y/N):')
-        if answer not in ['Y', 'y']:
-            logger.info('Aborting clean...')
-            return
-
-    for container in containers:
-        logger.info(f'Removing {container.name}...')
-        container.clean()
-
-def env_build_container(args, settings):
-    for c in get_containers_by_selector(args.images, settings):
-        c.build(force=args.force)
-
 def ensure_container_exists(container):
     if not container.exists:
         print(f"Missing container '{container.name}'\n")
@@ -341,202 +189,6 @@ def checkstyle(args, settings):
             print(f"{filename}:exception while reading file:{e}")
 
 
-class Build(object):
-    def __init__(self, name, container, settings, commit=None):
-        self.name = name
-        self.container = container
-        self.settings = settings
-        self.commit = commit
-
-    @property
-    def build_base_dir(self):
-        if self.commit is not None:
-            return os.path.join(self.settings.cache_dir, f'builds_{self.commit}')
-        return os.path.join(self.settings.cache_dir, f'builds')
-
-    @property
-    def build_dir(self):
-        return os.path.join(self.build_base_dir, self.container.name, self.name)
-
-    @property
-    def build_script(self):
-        return os.path.join(self.settings.build_scripts_dir, f"{self.name}.sh")
-
-    @property
-    def build_result_file(self):
-        return os.path.join(self.build_dir, "result.json")
-
-    @property
-    def exists(self):
-        return os.path.exists(self.build_result_file)
-
-    @property
-    def state(self):
-        if self.exists:
-            with open(self.build_result_file, "r") as f:
-                result = json.load(f)
-
-            if result["return_code"] == 0:
-                return "success"
-            return "failure"
-        else:
-            return "not executed"
-
-class CompilationTest(Build):
-    def __init__(self, name, container, settings, commit=None):
-        super(CompilationTest, self).__init__(name, container, settings, commit)
-
-    def build(self):
-        workdir = self.build_dir
-        build_script = self.build_script
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-
-        try:
-            return_code = self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=build_script,
-                                   cwd=workdir)
-        except KeyboardInterrupt:
-            return_code = -1
-
-        with open(self.build_result_file, "w") as f:
-            result = {
-                'return_code': return_code
-            }
-            json.dump(result, f)
-
-        return return_code == 0
-
-class RunTest(object):
-    def __init__(self, container, settings, ignore_lammps_commit=False):
-        self.container = container
-        self.settings = settings
-        self.ignore_lammps_commit = ignore_lammps_commit
-
-    @property
-    def build_base_dir(self):
-        if self.ignore_lammps_commit:
-            return os.path.join(self.settings.cache_dir, f'builds')
-        return os.path.join(self.settings.cache_dir, f'builds_{self.settings.current_lammps_commit}')
-
-    def get_build_dir(self, build_name):
-        return os.path.join(self.build_base_dir, self.container.name, build_name)
-
-    def get_build_script(self, build_name):
-        return os.path.join(self.settings.run_tests_scripts_dir, build_name, "build.sh")
-
-    def get_test_script(self, build_name):
-        return os.path.join(self.settings.run_tests_scripts_dir, build_name, "test.sh")
-
-    def build(self, build_name):
-        workdir = self.get_build_dir(build_name)
-        build_script = self.get_build_script(build_name)
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-        return self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=build_script,
-                                   cwd=workdir) == 0
-
-    def test(self, build_name):
-        workdir = self.get_build_dir(build_name)
-        test_script = self.get_test_script(build_name)
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-        return self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=test_script,
-                                   cwd=workdir) == 0
-
-class UnitTest(object):
-    def __init__(self, container, settings, ignore_lammps_commit=False):
-        self.container = container
-        self.settings = settings
-        self.ignore_lammps_commit = ignore_lammps_commit
-
-    @property
-    def build_base_dir(self):
-        if self.ignore_lammps_commit:
-            return os.path.join(self.settings.cache_dir, f'builds')
-        return os.path.join(self.settings.cache_dir, f'builds_{self.settings.current_lammps_commit}')
-
-    def get_build_dir(self, build_name):
-        return os.path.join(self.build_base_dir, self.container.name, build_name)
-
-    def get_build_script(self, build_name):
-        return os.path.join(self.settings.unit_tests_scripts_dir, build_name, "build.sh")
-
-    def get_test_script(self, build_name):
-        return os.path.join(self.settings.unit_tests_scripts_dir, build_name, "test.sh")
-
-    def build(self, build_name):
-        workdir = self.get_build_dir(build_name)
-        build_script = self.get_build_script(build_name)
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-        return self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=build_script,
-                                   cwd=workdir) == 0
-
-    def test(self, build_name):
-        workdir = self.get_build_dir(build_name)
-        test_script = self.get_test_script(build_name)
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-        return self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=test_script,
-                                   cwd=workdir) == 0
-
-class RegressionTest(object):
-    def __init__(self, container, settings, ignore_lammps_commit=False):
-        self.container = container
-        self.settings = settings
-        self.ignore_lammps_commit = ignore_lammps_commit
-
-    @property
-    def build_base_dir(self):
-        if self.ignore_lammps_commit:
-            return os.path.join(self.settings.cache_dir, f'builds')
-        return os.path.join(self.settings.cache_dir, f'builds_{self.settings.current_lammps_commit}')
-
-    def get_build_dir(self, build_name):
-        return os.path.join(self.build_base_dir, self.container.name, build_name)
-
-    def get_build_script(self, build_name):
-        return os.path.join(self.settings.regression_scripts_dir, build_name, "build.sh")
-
-    def get_test_script(self, build_name):
-        return os.path.join(self.settings.regression_scripts_dir, build_name, "test.sh")
-
-    def build(self, build_name):
-        workdir = self.get_build_dir(build_name)
-        build_script = self.get_build_script(build_name)
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-        return self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=build_script,
-                                   cwd=workdir) == 0
-
-    def test(self, build_name):
-        workdir = self.get_build_dir(build_name)
-        test_script = self.get_test_script(build_name)
-
-        os.makedirs(workdir, exist_ok=True)
-        LAMMPS_DIR = self.settings.lammps_dir
-        BUILD_SCRIPTS_DIR = self.settings.build_scripts_dir
-        return self.container.exec(options=['-B', f'{LAMMPS_DIR}/:{LAMMPS_DIR}/', '-B', f'{BUILD_SCRIPTS_DIR}/:{BUILD_SCRIPTS_DIR}/'],
-                                   command=test_script,
-                                   cwd=workdir) == 0
 
 def expand_selected_config_and_builds(selected_builds, settings):
     builds = {}
@@ -689,147 +341,6 @@ def unit_test(args, settings):
                     print(f"Run unit tests of '{build_name}' on '{container.name}' FAILED!")
                     sys.exit(1)
 
-def env_list(args, settings):
-    containers = get_containers(settings)
-    print("local")
-    for c in containers:
-        print(c.name)
-
-def env_status(args, settings):
-    containers = get_containers(settings)
-    print(f" {state_icon('success')} local")
-    for c in containers:
-        if c.exists:
-            icon = state_icon("success")
-        else:
-            icon = state_icon("")
-
-        print(f" {icon} {c.name}")
-
-def config_list(args, settings):
-    for c in get_configurations(settings):
-        print(c.name)
-
-def build_list(args, settings):
-    try:
-        configs = get_configurations_by_selector(args.config, settings)
-        for config in configs:
-            if hasattr(config, "builds"):
-                for build in config.builds:
-                    print(f"{config.name}/{build}")
-    except FileNotFoundError:
-        print(f"Configuration with name '{args.config}' does not exist!")
-        sys.exit(1)
-
-def state_icon(state):
-    if state == "success":
-        return "✅"
-    elif state == "failed":
-        return "❌"
-    else:
-        return "⭕"
-
-def print_build_status(config, build):
-    icon = state_icon(build.state)
-    commit_info = ""
-
-    if build.commit is not None:
-        commit_info = f" ({build.commit[:8]:8s})"
-
-    print(f" {icon} {config.name}/{build.name}{commit_info}")
-
-def build_status(args, settings):
-    try:
-        configs = get_configurations_by_selector(args.config, settings)
-
-        for config in configs:
-            container = get_container(config.container_image, settings)
-
-            if hasattr(config, "builds"):
-                for build in config.builds:
-                    if args.ignore_commit:
-                        compTest = CompilationTest(build, container, settings)
-                    else:
-                        compTest = CompilationTest(build, container, settings, settings.current_lammps_commit)
-                    print_build_status(config, compTest)
-    except FileNotFoundError:
-        print(f"Configuration with name '{args.config}' does not exist!")
-        sys.exit(1)
-
-def reg_list(args, settings):
-    examples_dir = os.path.join(settings.lammps_dir, 'examples')
-    testcase_count = 0
-    run_count = 0
-    mpi_run_count = 0
-    for test, scripts, logfiles in discover_tests(examples_dir):
-        example = LAMMPSExample(test, scripts, logfiles)
-        if len(example.testcases) > 0:
-            print(example.name)
-            for testcase, runs in example.testcases.items():
-                print(" -", testcase)
-                testcase_count += 1
-                for run in runs:
-                    if "MPI" in str(run):
-                        mpi_run_count += 1
-                    print("    *", run)
-                    run_count += 1
-    print("Total testcases:", testcase_count)
-    print("Total runs:", run_count)
-    print("Total MPI runs:", mpi_run_count)
-
-def reg_create_yaml(args, settings):
-    examples_dir = os.path.join(settings.lammps_dir, 'examples')
-    for test, scripts, logfiles in discover_tests(examples_dir):
-        example = LAMMPSExample(test, scripts, logfiles)
-        if len(example.testcases) > 0:
-            example.save_config()
-
-def init_env_command(parser):
-    subparsers = parser.add_subparsers(help='sub-command help')
-
-    elist = subparsers.add_parser('list', help='list all test environments')
-    elist.set_defaults(func=env_list)
-
-    status = subparsers.add_parser('status', help='show build status of all test environments')
-    status.set_defaults(func=env_status)
-
-    build = subparsers.add_parser('build', help='build container image(s)')
-    build.add_argument('images', metavar='image_name', nargs='+', help='container image names')
-    build.add_argument('-f', '--force', default=False, action='store_true', help="Force rebuild")
-    build.set_defaults(func=env_build_container)
-
-    clean = subparsers.add_parser('clean', help='remove container image(s)')
-    clean.add_argument('images', metavar='image_name', nargs='+', help='container image names')
-    clean.add_argument('-f', '--force', default=False, action='store_true', help="Clean without asking")
-    clean.set_defaults(func=env_clean)
-
-def init_config_command(parser):
-    subparsers = parser.add_subparsers(help='sub-command help')
-
-    clist = subparsers.add_parser('list', help='list all configurations')
-    clist.set_defaults(func=config_list)
-
-def init_build_command(parser):
-    subparsers = parser.add_subparsers(help='sub-command help')
-
-    blist = subparsers.add_parser('list', help='list all configurations')
-    blist.add_argument('config', metavar='config', default=['ALL'], help='name of configuration', nargs='*')
-    blist.set_defaults(func=build_list)
-
-    status = subparsers.add_parser('status', help='show build status')
-    status.add_argument('config', metavar='config', default=['ALL'], help='name of configuration', nargs='*')
-    status.add_argument('--ignore-commit', default=False, action='store_true', help='Ignore commit and do not create SHA specific build folder')
-    status.set_defaults(func=build_status)
-
-def init_reg_command(parser):
-    subparsers = parser.add_subparsers(help='sub-command help')
-
-    blist = subparsers.add_parser('list', help='list all configurations')
-    blist.add_argument('config', metavar='config', default=['ALL'], help='name of configuration', nargs='*')
-    blist.set_defaults(func=reg_list)
-
-    create_yaml = subparsers.add_parser('create_yaml', help='generate yaml files for all cases')
-    create_yaml.set_defaults(func=reg_create_yaml)
 
 def main():
     s = Settings()
